apps/geocore-tileserver/requirements.txt
APScheduler==3.6.3
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.4
django-appconf==1.0.4
django-cors-headers==3.2.1
django-filter==1.1.0
django-tilestache==2.0.9
djangorestframework==3.11.0
djangorestframework-filters==0.11.1
djangorestframework-gis==0.15
GDAL==2.2.2
gunicorn==20.0.4
idna==2.9
jsonfield==3.0.0
mapbox-vector-tile==1.2.0
ModestMaps==1.4.7
Pillow==7.1.1
psycopg2-binary==2.8.4
pyclipper==1.1.0.post3
pylru==1.2.0
pytz==2019.3
redis==3.4.1
requests==2.23.0
Shapely==1.7.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
TileStache==1.51.14
tzlocal==2.0.0
urllib3==1.25.8
Werkzeug==1.0.1